+++ packages/larva-js/src/interface/VideoShowcase/VideoShowcase.js
@@ -197,6 +197,10 @@ export default class VideoShowcase {
 		if ( 'jwplayer' === type ) {
 			return `https://content.jwplatform.com/feeds/${id}.json`;
 		}
+
+		if ( 'twitch' === type ) {
+			return `https://player.twitch.tv/?video=${id}`;
+		}
 	}
 
 	/**
@@ -207,6 +211,16 @@ export default class VideoShowcase {
 	playYoutube( youtubeUrl ) {
 		this.playerUI.iframe.removeAttribute( 'hidden' );
 		this.playerUI.iframe.setAttribute( 'src', `${youtubeUrl}?rel=0&autoplay=1&showinfo=0&controls=2&rel=0&modestbranding=0` );
+	}
+
+	/**
+	 * Remove hidden attribute from the iframe and set the src.
+	 *
+	 * @param {string} twitchUrl - A Twitch embed URL from returnUrl.
+	 */
+	 playTwitch( twitchUrl ) {
+		this.playerUI.iframe.removeAttribute( 'hidden' );
+		this.playerUI.iframe.setAttribute( 'src', `${twitchUrl}&autoplay=true&parent=${window.location.hostname}` );
 	}
 
 	/**
@@ -265,7 +279,7 @@ export default class VideoShowcase {
 	 * A wrapper function to conditonally play videos according to their type.
 	 *
 	 * @param {string} id - Youtube or JWplayer ID, should be from this.state.videoID, e.g. f1FX5wvC3DA
-	 * @param {string} type - "youtube" or "jwplayer"
+	 * @param {string} type - "youtube" or "jwplayer" or "twitch"
 	 */
 	playVideo( id, type ) {
 		let url = this.returnUrl( id, type );
@@ -276,6 +290,10 @@ export default class VideoShowcase {
 
 		if ( 'jwplayer' === type ) {
 			this.playJW( url );
+		}
+
+		if ( 'twitch' === type ) {
+			this.playTwitch( url );
 		}
 	}